requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.1
docstring_parser==0.16
drf-yasg==1.21.10
exceptiongroup==1.3.0
flatbuffers==25.2.10
fonttools==4.58.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-aiplatform==1.94.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.16.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jax==0.6.1
jaxlib==0.6.1
jiter==0.10.0
kiwisolver==1.4.8
matplotlib==3.10.3
ml_dtypes==0.5.1
openai==1.82.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
sentencepiece==0.2.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
websockets==15.0.1
whitenoise==6.6.0
